src/requirements.txt
PyYAML==6.0.1
Jinja2==3.1.3
ttast==0.3.2
jsonpatch==1.33
